+++ instrumentation/opentelemetry-instrumentation-wsgi/src/opentelemetry/instrumentation/wsgi/package.py
@@ -14,3 +14,5 @@
 
 
 _instruments = tuple()
+
+_supports_metrics = True